Beam functions optionally return a beam object
The beam functions can now optionally return a beam object instead of the beam id. This was not set
a default behavior for compatibility with current apps using this library. In the future we may add
deprecation for getting a beam id instead of an object.

The server does not return 'files' in the JSON returned when initiating a beam, which is why we had
to change Beam.from_json